requirements.txt
flask>=3.0
flask-cors>=4.0
groq>=0.9
gunicorn>=22.0
python-dotenv>=1.0
pdfminer.six>=20221105
numpy>=1.26
transformers>=4.40
sentence-transformers>=3.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.3
